Break resolution when other middleware present
This tweaks the movie schema to demonstrate that the `absinthe_auth`
middleware does not play nicely when other middleware is present. This
is because `absinthe` strips the default middleware for fields that have
provided middleware but still need the MapGet default. `absinthe_auth`
was adding this back in, but making the assumption that it should not if
there is other middleware down the line. The following fix is to be more
precise in checking that there are no resolvers in the middleware before
jamming the value back in.