Play nice with other middleware for default fields (#10)
* Break resolution when other middleware present

This tweaks the movie schema to demonstrate that the `absinthe_auth`
middleware does not play nicely when other middleware is present. This
is because `absinthe` strips the default middleware for fields that have
provided middleware but still need the MapGet default. `absinthe_auth`
was adding this back in, but making the assumption that it should not if
there is other middleware down the line. The following fix is to be more
precise in checking that there are no resolvers in the middleware before
jamming the value back in.

* Check for specific resolvers before defaulting

This checks for the Absinthe.Resolution or default middleware in order
to jam back in the default if it has been stripped. We do not want to
insert the default middleware, however, if we have a function being used
for resolution for a field.

* Set MIX_ENV to test for CI to check PRs